rename name -> participant-name
