+++ src/domain/participant/__tests__/check-email-already-exists.test.ts
@@ -1,7 +1,7 @@
 import { CheckEmailAlreadyExists } from "../check-email-already-exists"
 import { Email } from "../email"
 import { IParticipantRepository } from "../interface/participant-repository"
-import { Name } from "../name"
+import { ParticipantName } from "../participant-name"
 import { Participant } from "../participant"
 import { ParticipantId } from "../participant-id"
 
@@ -18,7 +18,7 @@ describe("CheckEmailAlreadyExists", () => {
   it("与えられたメールアドレスを持つ参加者が1人以上存在する場合 true を返す", async () => {
     participantRepositoryMock.getParticipantsByEmail.mockResolvedValue([
       Participant.reconstruct(ParticipantId.reconstruct("test"), {
-        name: Name.reconstruct("test"),
+        name: ParticipantName.reconstruct("test"),
         email: Email.reconstruct("test@example.com"),
       }),
     ])
