+++ src/domain/participant/participant-factory.ts
@@ -1,7 +1,7 @@
 import { DomainService } from "../shared/domainService"
 import { Participant } from "./participant"
 import { ParticipantId } from "./participant-id"
-import { Name } from "./name"
+import { ParticipantName } from "./participant-name"
 import { Email } from "./email"
 import { CheckEmailAlreadyExists } from "./check-email-already-exists"
 
@@ -17,7 +17,7 @@ export class ParticipantFactory extends DomainService<"participant-factory"> {
 
   async create({ name, email }: ParticipantFactoryProps): Promise<Participant> {
     const id = ParticipantId.create()
-    const nameValueObject = Name.create(name)
+    const nameValueObject = ParticipantName.create(name)
     const emailValueObject = await Email.create(
       email,
       this.checkEmailAlreadyExists
